+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/interop/Ksp2AdditionalSourceProvider.kt
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 package dev.zacsweers.metro.compiler.interop
 
-import com.google.devtools.ksp.impl.KSPCoreEnvironment
 import com.google.devtools.ksp.impl.KotlinSymbolProcessing
 import com.google.devtools.ksp.processing.KSPJvmConfig
 import com.google.devtools.ksp.processing.SymbolProcessorProvider
@@ -12,8 +11,6 @@ import dev.zacsweers.metro.compiler.test.JVM_TARGET
 import java.io.File
 import java.util.EnumSet
 import java.util.ServiceLoader
-import ksp.com.intellij.openapi.application.ApplicationManager as ShadedKspApplicationManager
-import ksp.com.intellij.openapi.util.Disposer as ShadedKspDisposer
 import org.jetbrains.kotlin.cli.common.messages.CompilerMessageSeverity
 import org.jetbrains.kotlin.cli.common.messages.MessageRenderer
 import org.jetbrains.kotlin.cli.common.messages.PrintingMessageCollector
@@ -168,8 +165,6 @@ class Ksp2AdditionalSourceProvider(testServices: TestServices) :
           .ifEmpty { EnumSet.of(CompilerMessageSeverity.ERROR, CompilerMessageSeverity.EXCEPTION) }
           .let { EnumSet.copyOf(it) }
       logger.reportAll(reportToCompilerSeverity)
-
-      clearKspLeaks()
     }
 
     val kotlinKspTestFiles =
@@ -177,19 +172,6 @@ class Ksp2AdditionalSourceProvider(testServices: TestServices) :
     val javaKspTestFiles =
       javaOutput.walkTopDown().filter { it.isFile }.map { it.toTestFile() }.toList()
     return kotlinKspTestFiles + javaKspTestFiles
-  }
-
-  /**
-   * KSP leaks its core environment because its CLI appears to be intended for single-shot use and
-   * stores stuff in ThreadLocals.
-   *
-   * The Gradle plugin doesn't seem to run into this because it runs all the tasks in an isolated
-   * classloader that dies between task runs.
-   */
-  private fun clearKspLeaks() {
-    KSPCoreEnvironment.instance_prop.remove()
-    // Doesn't _seem_ necessary but just in case, since it does appear to spin this up
-    ShadedKspApplicationManager.getApplication()?.let(ShadedKspDisposer::dispose)
   }
 
   private companion object {